refactor(front): move orglayout as a component
